backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18
pydantic==2.10.3
pydantic-settings==2.6.1
supabase==2.10.0
python-dotenv==1.0.1
reportlab==4.2.5
Pillow==11.0.0
httpx>=0.26,<0.28
